+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 * Add RTL/Jest testing for `FinePolicyDetail` component in `src/settings/FinePolicy`. Refs UICIRC-738.
 * Add RTL/Jest testing for `FixedDueDateScheduleManager` component in `src/settings/FixedDueDateSchedule`. Refs UICIRC-600.
 * Also support `circulation` `13.0`. Refs UICIRC-732.
+* Date format in preview should change depending on localization in `Staff slips`. Refs UICIRC-734.
 
 ## [6.0.0](https://github.com/folio-org/ui-circulation/tree/v6.0.0) (2021-09-30)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v5.1.1...v6.0.0)